[KOGITO-5682] Rename kogito events/messaging addons Apply suggestions from code review
