+++ addons/common/cloudevents/common/pom.xml
@@ -21,13 +21,13 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>kogito-addons-cloudevents</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents</name>
-  <description>CloudEvents Add-On Common Interfaces</description>
+  <artifactId>kogito-addons-messaging</artifactId>
+  <name>Kogito :: Add-Ons :: Messaging</name>
+  <description>Messaging Add-On Common Interfaces</description>
 
   <dependencies>
     <dependency>
