+++ addons/common/cloudevents/pom.xml
@@ -24,11 +24,11 @@
     <artifactId>kogito-addons-common-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-cloudevents-parent</artifactId>
+  <artifactId>kogito-addons-messaging-parent</artifactId>
   <packaging>pom</packaging>
-  <name>Kogito :: Add-Ons :: Cloud Events</name>
+  <name>Kogito :: Add-Ons :: Messaging</name>
 
-  <description>Kogito Cloud Events</description>
+  <description>Kogito Messaging parent</description>
 
   <modules>
     <module>common</module>
