+++ addons/common/cloudevents/utils/pom.xml
@@ -5,7 +5,7 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
