+++ addons/deprecated/kogito-addons-quarkus-cloudevents-multi/pom.xml
@@ -1,0 +1,22 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.kie.kogito</groupId>
+    <artifactId>kogito-addons-deprecated-parent</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>kogito-addons-quarkus-cloudevents-multi</artifactId>
+  <name>Kogito :: Add-Ons :: Deprecated CloudEvents :: Multi channel :: Quarkus</name>
+  <description>The library kogito-addons-quarkus-cloudevents-multi is deprecated and will be removed in a future release!. The functionality provided by this addon was merged into
+ *             kogito-addons-quarkus-messaging</description>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>kogito-addons-quarkus-messaging</artifactId>
+    </dependency>
+  </dependencies>
+
+</project>
