+++ addons/deprecated/kogito-addons-quarkus-cloudevents/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>kogito-addons-deprecated-parent</artifactId>
+    <groupId>org.kie.kogito</groupId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>kogito-addons-quarkus-cloudevents</artifactId>
+  <name>Kogito :: Add-Ons :: Deprecated :: CloudEvents :: Quarkus</name>
+  <description>The library kogito-addons-quarkus-cloudevents is deprecated and will be removed in a future release!. The new name for this addon is kogito-addons-quarkus-messaging</description>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>kogito-addons-quarkus-messaging</artifactId>
+    </dependency>
+  </dependencies>
+</project>
