+++ addons/deprecated/kogito-addons-quarkus-events-smallrye/pom.xml
@@ -1,0 +1,21 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.kie.kogito</groupId>
+    <artifactId>kogito-addons-deprecated-parent</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>kogito-addons-quarkus-events-smallrye</artifactId>
+  <name>Kogito :: Add-Ons :: Deprecated:: Events :: Smallrye Messaging</name>
+  <description>The library kogito-addons-quarkus-events-smallrye is deprecated and will be removed in a future release!. The new name for this addon is kogito-addons-quarkus-events-process</description>
+
+
+  <dependencies>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>kogito-addons-quarkus-events-process</artifactId>
+    </dependency>
+  </dependencies>
+</project>
