+++ addons/deprecated/kogito-addons-springboot-cloudevents/src/main/java/org/kie/kogito/addon/cloudevents/springboot/deprecated/DeprecatedCloudEventsSpringBoot.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.addon.cloudevents.springboot.deprecated;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+/**
+ * @deprecated The library kogito-addons-springboot-cloudevents is deprecated and will be removed in a future release!. The new name for this addon is kogito-addons-springboot-messaging
+ */
+
+@Deprecated
+@Component
+public class DeprecatedCloudEventsSpringBoot {
+
+    public static final Logger LOGGER = LoggerFactory.getLogger(DeprecatedCloudEventsSpringBoot.class);
+
+    public DeprecatedCloudEventsSpringBoot() {
+        LOGGER.warn(
+                "@deprecated The library kogito-addons-springboot-cloudevents is deprecated and will be removed in a future release!. The new name for this addon is kogito-addons-springboot-messaging");
+    }
+
+}
