+++ addons/deprecated/kogito-addons-springboot-events-kafka/pom.xml
@@ -1,0 +1,20 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.kie.kogito</groupId>
+    <artifactId>kogito-addons-deprecated-parent</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <artifactId>kogito-addons-springboot-events-kafka</artifactId>
+  <name>Kogito :: Add-Ons :: Events :: SprintBoot :: Kafka</name>
+  <description>Kogito Events based on Spring Boot Add-On</description>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>The library kogito-addons-springboot-events-kafka is deprecated and will be removed in a future release!. The new name for this addon is kogito-addons-springboot-events-process-kafka</artifactId>
+    </dependency>
+  </dependencies>
+</project>
