+++ quarkus/addons/cloudevents/common/pom.xml
@@ -4,17 +4,17 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-quarkus-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-cloudevents-common</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents :: Quarkus</name>
-  <description>Common classes for Kogito CloudEvents AddOn (Quarkus)</description>
+  <artifactId>kogito-addons-quarkus-messaging-common</artifactId>
+  <name>Kogito :: Add-Ons :: Messages :: Quarkus:: Common</name>
+  <description>Common classes for Kogito messaging Add-On (Quarkus)</description>
 
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-cloudevents</artifactId>
+      <artifactId>kogito-addons-messaging</artifactId>
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
