+++ quarkus/addons/cloudevents/implementation/pom.xml
@@ -3,20 +3,20 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>kogito-addons-quarkus-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-messaging-parent</artifactId>
     <groupId>org.kie.kogito</groupId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>kogito-addons-quarkus-cloudevents</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents :: Quarkus</name>
-  <description>Kogito Events handling for CloudEvents AddOn (Quarkus)</description>
+  <artifactId>kogito-addons-quarkus-messaging</artifactId>
+  <name>Kogito :: Add-Ons :: Messaging :: Quarkus</name>
+  <description>Kogito Events handling for BPMN messages Add-On (Quarkus)</description>
 
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-quarkus-cloudevents-common</artifactId>
+      <artifactId>kogito-addons-quarkus-messaging-common</artifactId>
     </dependency>
     <dependency>
       <groupId>io.cloudevents</groupId>
