+++ quarkus/addons/cloudevents/integration-tests/pom.xml
@@ -4,12 +4,12 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-quarkus-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-cloudevents-it</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents :: Quarkus (ITs)</name>
-  <description>Integration tests for for CloudEvents AddOn (Quarkus)</description>
+  <artifactId>kogito-addons-quarkus-messaging-it</artifactId>
+  <name>Kogito :: Add-Ons :: Messaging :: Quarkus (ITs)</name>
+  <description>Integration tests for for Messaging Add-On (Quarkus)</description>
 
   <dependencyManagement>
     <dependencies>
@@ -30,9 +30,9 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-quarkus-cloudevents</artifactId>
+      <artifactId>kogito-addons-quarkus-messaging</artifactId>
     </dependency>
-    <dependency>
+     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-smallrye-reactive-messaging-kafka</artifactId>
     </dependency>
