--- quarkus/addons/cloudevents/multi/src/main/resources/META-INF/kogito.addon
@@ -1,1 +1,0 @@
-cloudevents
