+++ quarkus/addons/cloudevents/pom.xml
@@ -7,16 +7,15 @@
     <artifactId>kogito-addons-quarkus-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-cloudevents-parent</artifactId>
+  <artifactId>kogito-addons-quarkus-messaging-parent</artifactId>
   <packaging>pom</packaging>
-  <name>Kogito :: Add-Ons :: Quarkus :: Cloud Events</name>
+  <name>Kogito :: Add-Ons :: Quarkus :: Messaging</name>
 
-  <description>Kogito Cloud Events</description>
+  <description>Kogito Quarkus Messaging Parent</description>
 
   <modules>
     <module>common</module>
     <module>implementation</module>
-    <module>multi</module>
   </modules>
 
   <profiles>
