+++ quarkus/addons/events/smallrye/pom.xml
@@ -7,9 +7,9 @@
     <artifactId>kogito-addons-quarkus-events-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-quarkus-events-smallrye</artifactId>
-  <name>Kogito :: Add-Ons :: Events :: Smallrye Messaging</name>
-  <description>Kogito Events based on Smallrye Reactive Messaging AddOn</description>
+  <artifactId>kogito-addons-quarkus-events-process</artifactId>
+  <name>Kogito :: Add-Ons :: Events :: Process</name>
+  <description>Kogito AddOn for runtime process events </description>
 
   <dependencies>
     <dependency>
