+++ springboot/addons/cloudevents/implementation/pom.xml
@@ -21,17 +21,17 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-springboot-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-springboot-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-springboot-cloudevents</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents :: Spring Boot</name>
-  <description>Kogito Events handling for CloudEvents AddOn (Spring Boot)</description>
+  <artifactId>kogito-addons-springboot-messaging</artifactId>
+  <name>Kogito :: Add-Ons :: Messaging :: Spring Boot</name>
+  <description>Kogito Events handling for Messaging Add-On (Spring Boot)</description>
 
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-cloudevents</artifactId>
+      <artifactId>kogito-addons-messaging</artifactId>
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
