+++ springboot/addons/cloudevents/integration-tests/pom.xml
@@ -7,12 +7,12 @@
 
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-springboot-cloudevents-parent</artifactId>
+    <artifactId>kogito-addons-springboot-messaging-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-springboot-cloudevents-it</artifactId>
-  <name>Kogito :: Add-Ons :: CloudEvents :: Spring Boot (ITs)</name>
-  <description>Integration tests for for CloudEvents AddOn (Spring Boot)</description>
+  <artifactId>kogito-addons-springboot-messaging-it</artifactId>
+  <name>Kogito :: Add-Ons :: Messaging :: Spring Boot (ITs)</name>
+  <description>Integration tests for for Messages AddOn (Spring Boot)</description>
 
   <dependencies>
     <!-- All the dependency of this module are here just to make Maven reactor aware that invoker tests need them -->
@@ -46,7 +46,7 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-springboot-cloudevents</artifactId>
+      <artifactId>kogito-addons-springboot-messaging</artifactId>
     </dependency>
   </dependencies>
