+++ springboot/addons/cloudevents/integration-tests/src/it/cloudevents-spring-boot-addon-it-invoked/pom.xml
@@ -25,7 +25,7 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>kogito-addons-springboot-cloudevents</artifactId>
+      <artifactId>kogito-addons-springboot-messaging</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
