+++ springboot/addons/events/kafka/pom.xml
@@ -7,9 +7,9 @@
     <artifactId>kogito-addons-springboot-events-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>kogito-addons-springboot-events-kafka</artifactId>
-  <name>Kogito :: Add-Ons :: Events :: SprintBoot :: Kafka</name>
-  <description>Kogito Events based on Spring Boot Add-On</description>
+  <artifactId>kogito-addons-springboot-events-process-kafka</artifactId>
+  <name>Kogito :: Add-Ons :: Events :: SprintBoot :: Process :: kafka</name>
+  <description>Kogito Kafka Events for process based on Spring Boot Add-On</description>
 
   <dependencies>
     <dependency>
